+++ apps/governance/src/toasts-manager.tsx
@@ -1,11 +1,11 @@
 import { ToastsContainer, useToasts } from '@vegaprotocol/ui-toolkit';
-import { useUpdateNetworkParametersToasts } from '@vegaprotocol/proposals';
-import { useVegaTransactionToasts } from '@vegaprotocol/web3';
-import { useEthereumTransactionToasts } from '@vegaprotocol/web3';
-import { useEthereumWithdrawApprovalsToasts } from '@vegaprotocol/web3';
+import {
+  useEthereumTransactionToasts,
+  useEthereumWithdrawApprovalsToasts,
+  useVegaTransactionToasts,
+} from '@vegaprotocol/web3';
 
 export const ToastsManager = () => {
-  useUpdateNetworkParametersToasts();
   useVegaTransactionToasts();
   useEthereumTransactionToasts();
   useEthereumWithdrawApprovalsToasts();
